requirements.txt
aiohttp==3.7.4.post0
appnope==0.1.0
async-timeout==3.0.1
attrs==21.2.0
backcall==0.2.0
base58==2.1.0
bitarray==1.2.2
bleach==3.3.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
crytic-compile==0.1.9
cytoolz==0.11.0
dateparser==1.0.0
decorator==4.4.2
discord==1.0.1
discord.py==1.7.2
docutils==0.16
eth-abi==2.1.1
eth-account==0.5.4
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.9.5
hexbytes==0.2.1
idna==2.10
ipfshttpclient==0.7.0a1
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
jsonschema==3.2.0
keyring==22.3.0
lru-dict==1.1.7
m2r==0.2.1
mistune==0.8.4
multiaddr==0.0.9
multidict==5.1.0
netaddr==0.8.0
numpy==1.20.2
packaging==20.9
pandas==1.2.4
parsimonious==0.8.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pkginfo==1.7.0
prettytable==1.0.1
prometheus-api-client==0.4.2
prompt-toolkit==3.0.7
protobuf==3.12.4
ptyprocess==0.6.0
pycoingecko==2.1.0
pycryptodome==3.10.1
Pygments==2.7.0
pyparsing==2.4.7
pyrsistent==0.17.3
pysha3==1.0.2
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2021.1
readme-renderer==29.0
regex==2021.4.4
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.4.0
rlp==2.0.1
six==1.15.0
slither-analyzer==0.6.13
solidity-flattener==0.2.2
toolz==0.11.1
tqdm==4.58.0
traitlets==5.0.4
twine==3.3.0
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.3
varint==1.0.2
wcwidth==0.2.5
web3==5.19.0
webencodings==0.5.1
websockets==8.1
yarl==1.6.3
